First commit of new examples
This likely needs some refactoring in terms of vanilla js and p5
examples. Also, where to store the libraries, models, and other shared
assets. With each example or in a separate folder? This repo
needs a lot more work (README, CONTRIBUTING, etc) at the moment all
I'm doing is copying over the folder right now.

This relates to:
https://github.com/ml5js/ml5-library/issues/76