Updates to the README
This is probably silly but somehow "ml5" looks right to me and "ML5"
does not. I'm insane, right? I do like the capital M on the logo.

The web editor examples still need to be created.

Also, there is further renaming to do based on:
https://github.com/ml5js/ml5-library/issues/98

Finally, I linked to the "releases" page on github for downloading
a stand-alone ml5 library file. Is that the best place? Or an npm page?