Console logging any errors.
Since the error argument is now required in any callbacks I'm adding some very crude error handling (just logging the error) to the console to demonstrate. I'm only putting this in the image classification / feature extractor examples as of now but it could end up in more places or we could eventually have a "error handling" example.